backend/requirements.txt
Django==3.0.3
python-ldap==3.2.0
django-auth-ldap==2.1.0
django-ldapdb==1.4.0
Jinja2==2.10.3
djangorestframework==3.11.0
django-cors-headers==3.2.1
PyYAML==5.3.1
uritemplate==3.0.1
drf-yasg==1.17.1
